+++ backend/src/main/java/com/stock/fund/domain/entity/Stock.java
@@ -2,10 +2,15 @@ package com.stock.fund.domain.entity;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
  * 股票领域实体
+ * <p>
+ * 数值精度规范:
+ * - totalShare/floatShare (股本): 2位小数
+ * - pe/pb (市盈率/市净率): 2位小数
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -15,10 +20,10 @@ public class Stock extends AggregateRoot<Long> {
     private String industry;      // 所属行业
     private String market;        // 市场
     private LocalDate listingDate; // 上市日期
-    private Double totalShare;    // 总股本
-    private Double floatShare;    // 流通股本
-    private Double pe;            // 市盈率
-    private Double pb;            // 市净率
+    private BigDecimal totalShare;    // 总股本 (2位小数)
+    private BigDecimal floatShare;    // 流通股本 (2位小数)
+    private BigDecimal pe;            // 市盈率 (2位小数)
+    private BigDecimal pb;            // 市净率 (2位小数)
 
     // 构造函数
     public Stock() {}
